Airflow/MWAA-2.5.1/requirements.txt
kubernetes==23.6.0
apache-airflow-providers-cncf-kubernetes==5.1.1